split training phases into two different program runs

+++ train.py
@@ -18,6 +18,7 @@ import numpy as np
 import tensorflow as tf
 import core.utils as utils
 from tqdm import tqdm
+import argparse
 from core.dataset import Dataset
 from core.yolov3 import YOLOV3
 from core.config import cfg
@@ -44,6 +45,8 @@ class YoloTrain(object):
         self.steps_per_period    = len(self.trainset)
         self.sess                = tf.Session(config=tf.ConfigProto(allow_soft_placement=True))
         self.output_dir          = cfg.TRAIN.OUTPUT_FOLDER
+        self.stage_1_ckpt        = cfg.TRAIN.STAGE_1_WEIGHT 
+        self.stage_2_ckpt        = cfg.TRAIN.STAGE_2_WEIGHT 
 
         if not os.path.exists(self.output_dir):
             os.makedirs(self.output_dir)
@@ -112,8 +115,7 @@ class YoloTrain(object):
         with tf.name_scope('loader_and_saver'):
             self.loader = tf.train.Saver(self.net_var)
 
-            # setted to max epoch to make sure, all ckpts are available when selecting the best performanting 
-            self.saver  = tf.train.Saver(tf.global_variables(), max_to_keep=self.first_stage_epochs)
+            self.saver  = tf.train.Saver(tf.global_variables(), max_to_keep=10)
 
         with tf.name_scope('summary'):
             tf.summary.scalar("learn_rate",      self.learn_rate)
@@ -192,8 +194,8 @@ class YoloTrain(object):
             
             train_epoch_loss = np.mean(train_epoch_loss)
             test_epoch_loss = np.mean(test_epoch_loss)
-
-            ckpt_file = self.output_dir + "/yolov3_test_loss=%.4f.ckpt" % test_epoch_loss
+            
+            ckpt_file = self.output_dir + self.stage_1_ckpt + "_test_loss=%.4f.ckpt" % test_epoch_loss
             log_time = time.strftime('%Y-%m-%d %H:%M:%S', time.localtime(time.time()))
             print("=> Epoch: %2d Time: %s Train loss: %.2f Test loss: %.2f Saving %s ..."
                             %(epoch, log_time, train_epoch_loss, test_epoch_loss, ckpt_file))
@@ -205,7 +207,6 @@ class YoloTrain(object):
 
         print(f"Done with first stage after {self.first_stage_epochs} epochs")
         print(f"Best test loss of {best_performance_test_loss} had checkpoint {best_performance_ckpt}")
-        print("Using this for second stage training!")
 
     def train_second_stage(self, best_performance_ckpt):
         self.sess.run(tf.global_variables_initializer())
@@ -267,7 +268,7 @@ class YoloTrain(object):
             train_epoch_loss = np.mean(train_epoch_loss)
             test_epoch_loss = np.mean(test_epoch_loss)
             
-            ckpt_file = self.output_dir + "/yolov3_test_loss=%.4f.ckpt" % test_epoch_loss
+            ckpt_file = self.output_dir + self.stage_2_ckpt + "_test_loss=%.4f.ckpt" % test_epoch_loss
             log_time = time.strftime('%Y-%m-%d %H:%M:%S', time.localtime(time.time()))
             print("=> Epoch: %2d Time: %s Train loss: %.2f Test loss: %.2f Saving %s ..."
                             %(epoch, log_time, train_epoch_loss, test_epoch_loss, ckpt_file))
@@ -278,17 +279,31 @@ class YoloTrain(object):
                 best_performance_test_loss = test_epoch_loss
                 best_performance_ckpt = ckpt_file + "-" + str(epoch)
 
-        print(f"Done with second stage after {self.first_stage_epochs} epochs")
+        print(f"Done with second stage after {self.second_stage_epochs} epochs")
         print(f"Best test loss of {best_performance_test_loss} had checkpoint {best_performance_ckpt}")
 
                             
 if __name__ == '__main__':
-    if os.path.exists(logdir): shutil.rmtree(logdir)
-    os.mkdir(logdir)
+    parser = argparse.ArgumentParser()
+    parser.add_argument("stage", type=int,
+                    help="Select the stage which shall be trained ('1' or '2')")
+    parser.add_argument("ckpt", type=str,
+                    help="Select the ckpt file which shall be trained in the second stage!")
+    args = parser.parse_args()
 
-    best_ckpt = YoloTrain().train_first_stage()
-    YoloTrain().train_second_stage(best_performance_ckpt=best_ckpt)
+    stage = int(args.stage)
+    first_stage_ckpt = args.ckpt
 
-
-
-
+    if stage is 1:
+        if os.path.exists(logdir): shutil.rmtree(logdir)
+        logdir = logdir + "stage1/"
+        os.mkdir(logdir)
+        YoloTrain().train_first_stage()
+    elif stage is 2:
+        logdir = logdir + "stage2/"
+        if os.path.exists(logdir): shutil.rmtree(logdir)
+        os.mkdir(logdir)
+        assert first_stage_ckpt is not None, "You want to train for second stage, but no ckpt is given!"
+        YoloTrain().train_second_stage(first_stage_ckpt)
+    else:
+        print(f"Stage {stage} not valid (must be 1 or 2), exiting...")
